requirements.txt
# Python依赖
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
openai>=1.0.0
markdown>=3.5.0
pillow>=10.0.0
loguru>=0.7.0

# 可选：增强功能
# anthropic>=0.18.0  # Claude API
# newspaper3k>=0.2.8  # 文章内容提取